+++ CHANGELOG.md
@@ -41,6 +41,7 @@ All notable changes to this project will be documented in this file.
   - Always delegate RouteAdd regardless of noUninstall flag
 - Telemetry
   - Include solana vote pubkey in global monitor metrics
+  - Run telemetry agent on pending and drained links
 
 ## [v0.8.1](https://github.com/malbeclabs/doublezero/compare/client/v0.8.0...client/v0.8.1) – 2025-01-12